fix parseInt int32 overflow and uint32 conversion of negative doubles
parseInt accumulated into a C int, so parseInt("ffffffff", 16) wrapped
to -1.  esprima parses hex literals with parseInt, which meant the
self-hosted compiler read 0xffffffff literals in its own source as -1.
Constant.getIntegerValue's 64-bit form then cast that -1.0 to uint32_t,
which saturates to 0 on arm64 (vs wrapping on x86), so stage2 emitted
nanboxing masks of 0x7fff00000000 instead of 0x7fffffffffff and stage3
binaries crashed on their first closure env access.  the bindings now
convert with wrapping (ToUint32) semantics as well.

with this the bootstrap reaches a fixed point: stage2 and stage3 are
byte-for-byte identical.

+++ ejs-llvm/constant.cpp
@@ -79,7 +79,11 @@ namespace ejsllvm {
         }
         else if (argc == 3 && EJSVAL_IS_NUMBER(args[2]) && ty->getPrimitiveSizeInBits() == 64) {
             uint64_t vhi = v;
-            uint32_t vlo = (uint32_t)EJSVAL_TO_NUMBER(args[2]);
+            // convert with ToUint32 (wrapping) semantics: a bare
+            // double->uint32_t cast of a negative value saturates to 0 on
+            // arm64, silently corrupting constants like 0xffffffff that
+            // reach us as -1
+            uint32_t vlo = (uint32_t)(int64_t)EJSVAL_TO_NUMBER(args[2]);
             return Value_new (llvm::Constant::getIntegerValue(ty, llvm::APInt(ty->getPrimitiveSizeInBits(), (int64_t)((vhi << 32) | vlo))));
         }
         else
