+++ node-llvm/constant.cpp
@@ -76,7 +76,9 @@ namespace jsllvm {
       // allow a 3 arg form for 64 bit ints:
       // constant = llvm.Constant.getIntegerValue types.int64, ch, cl
       uint64_t vhi = v;
-      uint32_t vlo = (uint32_t)info[2]->NumberValue(context).ToChecked();
+      // convert with ToUint32 (wrapping) semantics: a bare double->uint32_t
+      // cast of a negative value saturates to 0 on arm64
+      uint32_t vlo = (uint32_t)(int64_t)info[2]->NumberValue(context).ToChecked();
       result = Value::Create (llvm::Constant::getIntegerValue(ty, llvm::APInt(ty->getPrimitiveSizeInBits(), (int64_t)((vhi << 32) | vlo))));
     }
     else {
